Stored arguments in structure
Instead of storing arguments in a 2D dimensional array, I will be storing it in a structure for easier access. Process line is fully ready to send arguments to the exec_proc or the process that pipes the arguments. We just need to change those function parameters in order to accept structures.